requirements.txt
openai>=1.0.0
requests>=2.31.0
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0
python-dotenv>=1.0.0
huggingface_hub
